requirements.txt
streamlit
dnspython
validate-email-address
